Daesung

+++ backend/src/main/java/com/web/pocketmoney/service/wish/WishServiceImpl.java
@@ -5,22 +5,23 @@ import com.web.pocketmoney.dto.wish.WishDTO;
 import com.web.pocketmoney.dto.wish.WishPageRequestDTO;
 import com.web.pocketmoney.dto.wish.WishPageResultDTO;
 import com.web.pocketmoney.entity.board.Board;
+import com.web.pocketmoney.entity.board.BoardRepository;
 import com.web.pocketmoney.entity.user.User;
 import com.web.pocketmoney.entity.user.UserRepository;
 import com.web.pocketmoney.entity.wish.Wish;
 import com.web.pocketmoney.entity.wish.WishRepository;
+import com.web.pocketmoney.exception.CBoardNotFoundException;
 import com.web.pocketmoney.exception.CUserNotFoundException;
-import com.web.pocketmoney.exception.WishNotFoundException;
+import com.web.pocketmoney.exception.CWishNotFoundException;
+import com.web.pocketmoney.exception.CWishUserNotEqualCurrentException;
+import com.web.pocketmoney.exception.handler.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.hibernate.sql.Insert;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
 import java.util.function.Function;
 
 @Service
@@ -31,9 +32,13 @@ public class WishServiceImpl implements WishService{
     private final WishRepository wishRepository;
     private final UserRepository userRepository;
 
+    private final BoardRepository boardRepository;
+
 
     @Override
     public Long register(InsertWishDTO insertwishDTO) {
+        Board board = boardRepository.findById(insertwishDTO.getBoardId()).orElseThrow(()->
+                new CBoardNotFoundException("해당 게시물을 찾을 수 없습니다.", ErrorCode.FORBIDDEN));
 
         Wish wish = dtoToEntity(insertwishDTO);
         log.info("wishDTO : " + insertwishDTO);
@@ -43,11 +48,19 @@ public class WishServiceImpl implements WishService{
     }
 
     @Override
-    public void remove(Long id) {
+    public void remove(Long id, Long userId) {
+
+
+
         Wish wish = wishRepository.findById(id)
-                .orElseThrow(() -> new WishNotFoundException(
-                        "존재하지 않는 글이거나 관심글이 아닙니다."
+                .orElseThrow(() -> new CWishNotFoundException(
+                        "관심글이 존재하지 않습니다..", ErrorCode.FORBIDDEN
                 ));
+
+        log.info("wish : "+wish.getUserId().getId(), userId);
+        if(!(wish.getUserId().getId().equals(userId))){
+            throw new CWishUserNotEqualCurrentException("해당 관심글을 지울 수 있는 권한이 없습니다.", ErrorCode.FORBIDDEN);
+        }
         wishRepository.deleteById(wish.getId());
     }
